Fix dictionary

+++ source/qb/QbRecompiler.cpp
@@ -978,10 +978,10 @@ std::vector<u8> QbRecompiler::compile(std::string &source, ChecksumDictionary &s
                 bytes.push_back(((checksum) >> 16) & 0xFF);
                 bytes.push_back(((checksum) >> 24) & 0xFF);
 
-                unsigned long reversed = ((unsigned long)(checksum & 0xFF) << 24) |
-                                         ((unsigned long)((checksum >> 8) & 0xFF) << 16) |
-                                         ((unsigned long)((checksum >> 16) & 0xFF) << 8) |
-                                         ((unsigned long)((checksum >> 24) & 0xFF));
+                uint32_t reversed = ((checksum & 0xFF) << 24) |
+                                         (((checksum >> 8) & 0xFF) << 16) |
+                                         (((checksum >> 16) & 0xFF) << 8) |
+                                         (((checksum >> 24) & 0xFF));
 
                 symbols.populate(reversed, checksum_name);
